clean, refactor, add more object dependency checks
