+++ src/common/inventory.ts
@@ -13,12 +13,20 @@ import {
   minAPI,
   techAdoptionRules,
   dependencyRules,
-  dataModels
-
+  dataModels,
+  standardNamespaces
 } from './rules';
 import {
-  loggit
+  addKeyToObject,
+  getValue,
+  setValue,
+  incrementValue
+}
+from './JSONUtilities'
+import {
+  Loggit
 } from './logger';
+import { PRIORITY_BELOW_NORMAL } from 'constants';
 
 
 export class packageInventory {
@@ -36,8 +44,8 @@ export class packageInventory {
 
 
   public constructor() {
-    this.loggit = new loggit('isvtePlugin:PackageInventory');
-    this.loggit.loggit('Creating New Package Inventory');
+    this.loggit = new Loggit('isvtePlugin:PackageInventory');
+    this.loggit.logLine('Creating New Package Inventory');
   };
 
   public setMinAPI(newAPIVersion) {
@@ -45,25 +53,25 @@ export class packageInventory {
   };
 
   public setMetadata(md) {
-    this.loggit.loggit('Setting Metadata');
+    this.loggit.logLine('Setting Metadata');
     this._mdInv = md;
   };
 
 
-  private checkCondition(cond) {
-    this.loggit.loggit('Checking Condition:' + JSON.stringify(cond));
+  private checkCondition(cond ) {
+    this.loggit.logLine('Checking Condition:' + cond.metadataType + ' ' + cond.operator);
     let response = {
       conditionPass: false,
       passItems: [],
       failItems: []
     };
     if (!this.isConditionValid(cond)) {
-      this.loggit.loggit('Condition is not valid. Returning false')
+      this.loggit.logLine('Condition is not valid. Returning false')
       return response;
     }
       let mdTypeCount = this.getCountByMetadataType(cond.metadataType);
       if (mdTypeCount.length == 0) {
-        this.loggit.loggit('Checking Condition. empty response receieved. We should not end up here');
+        this.loggit.logLine('Checking Condition. empty response receieved. We should not end up here');
         mdTypeCount.push({
           property: cond.metadataType,
           value: -1
@@ -73,24 +81,28 @@ export class packageInventory {
       if (Array.isArray(cond.operand)) {
         for (let i = 0; i < cond.operand.length; i++) {
           if (cond.operand[i] === 'minAPI') {
-            this.loggit.loggit('Replacing minAPI placeholder with ' + this._minAPI);
+            this.loggit.logLine('Replacing minAPI placeholder with ' + this._minAPI);
             cond.operand[i] = this._minAPI;
           }
         }
       }
       else {
         if (cond.operand === 'minAPI') {
-          this.loggit.loggit('Replacing minAPI placeholder with ' + this._minAPI);
+          this.loggit.logLine('Replacing minAPI placeholder with ' + this._minAPI);
 
           cond.operand = this._minAPI;
         }
       }
+      let showDetails = false;
+      if (cond.showDetails) {
+        showDetails = true;
+      }
 
       for (var itemCount of mdTypeCount) {
-        this.loggit.loggit('Validating item:' + itemCount.property);
-        this.loggit.loggit('  Value:' + itemCount.value);
-        this.loggit.loggit('  Operator:' + cond.operator);
-        this.loggit.loggit('  CompareTo:' + cond.operand);
+        this.loggit.logLine('Validating item:' + itemCount.property);
+        this.loggit.logLine('  Value:' + itemCount.value);
+        this.loggit.logLine('  Operator:' + cond.operator);
+        this.loggit.logLine('  CompareTo:' + cond.operand);
         let itemPass = false;
         switch (cond.operator) {
           case 'always':
@@ -129,16 +141,16 @@ export class packageInventory {
             break; 
         }
         if (itemPass) {
-          this.loggit.loggit('    Condition Passed');
+          this.loggit.logLine('    Condition Passed');
           response.conditionPass = true;
-          response.passItems.push(itemCount.property);
+          response.passItems.push({name:itemCount.property,display:showDetails});
         } else {
-          this.loggit.loggit('    Condition Failed');
-          response.failItems.push(itemCount.property);
+          this.loggit.logLine('    Condition Failed');
+          response.failItems.push({name:itemCount.property,display:showDetails});
         }
       }
       if (cond.conditionOr != undefined) {
-        this.loggit.loggit('Checking OR condition');
+        this.loggit.logLine('Checking OR condition');
         //By default, don't process the OR if the condition has already passed
         if (cond.conditionOr.processAlways == true || !response.conditionPass) {
           let orResponse = this.checkCondition(cond.conditionOr);
@@ -149,24 +161,48 @@ export class packageInventory {
         }
       }
       if (cond.conditionAnd != undefined) {
-        this.loggit.loggit('Checking AND condition');
+        this.loggit.logLine('Checking AND condition');
         //By default, don't process the AND if the condition is already false
         if (cond.conditionAnd.processAlways == true || response.conditionPass) {
           let andResponse = this.checkCondition(cond.conditionAnd);
           if (cond.conditionAnd.conditionPerItem) {
-            this.loggit.loggit('Checking condition per item in the AND');
-            this.loggit.loggit('Core Condition passes: ' + JSON.stringify(response.passItems));
-            this.loggit.loggit('And Condition passes: ' + JSON.stringify(andResponse.passItems));
+            this.loggit.logLine('Checking condition per item in the AND');
+            this.loggit.logLine('Core Condition passes: ' + JSON.stringify(response.passItems));
+            this.loggit.logLine('And Condition passes: ' + JSON.stringify(andResponse.passItems));
             //Condition passes if the same property passes in both this condition and in conditionAnd
-            response.passItems = response.passItems.filter(item => andResponse.passItems.includes(item));
-            this.loggit.loggit('Intersection of core and and Conditions: ' + JSON.stringify(response.passItems));
+            let tmpItems = [];
+            for (const i of response.passItems) {
+              for (const ai of andResponse.passItems) {
+                if (i.name == ai.name) {
+                  tmpItems.push ({name: i.name, display: i.display || ai.display});
+                  break;
+                }
+              }
+            }
+            response.passItems = [...tmpItems];
+            //response.passItems = response.passItems.filter(item =>andResponse.passItems.includes(item));
+            this.loggit.logLine('Intersection of core and and Conditions: ' + JSON.stringify(response.passItems));
             //Logic correct here? Fail will return items that do not pass both conditions. Items that pass one condition but not the other are lost.
-            response.failItems = response.failItems.filter(item => andResponse.failItems.includes(item));
+            //response.failItems = response.failItems.filter(item => andResponse.failItems.includes(item));
+            tmpItems = [];
+            for (const i of response.failItems) {
+              for (const ai of andResponse.failItems) {
+                if (i.name == ai.name) {
+                  tmpItems.push ({name: i.name, display: i.display || ai.display});
+                  break;
+                }
+              }
+            }
+            response.failItems = [...tmpItems];
             response.conditionPass = response.passItems.length > 0;
           } else {
+            this.loggit.logLine('ANDing the conditions');
+            this.loggit.logJSON(response);
+            this.loggit.logJSON(andResponse);
             response.conditionPass = response.conditionPass && andResponse.conditionPass;
             response.passItems.push(...andResponse.passItems);
             response.failItems.push(...andResponse.failItems);
+            this.loggit.logJSON(response);
           }
         }
       }
@@ -175,7 +211,7 @@ export class packageInventory {
   }
 
   private isConditionValid(cond) {
-    this.loggit.loggit('Checking validity of Condition:' + JSON.stringify(cond));
+    this.loggit.logLine('Checking validity of Condition:')// + JSON.stringify(cond));
     let isValid = true;
     let validOperators = ['always', 'never', 'exists', 'notexists', 'null', 'gt', 'gte', 'lt', 'lte', 'eq','between'];
     let operatorsNeedOperand = ['gt', 'gte', 'lt', 'lte', 'eq'];
@@ -184,23 +220,23 @@ export class packageInventory {
     if (cond['expiration'] != undefined) {
       const now = new Date().toJSON();
       if (cond.expiration < now) {
-        this.loggit.loggit('Condition has expired. Expiration:' + cond.expiration);
+        this.loggit.logLine('Condition has expired. Expiration:' + cond.expiration);
         isValid = false;
       }
     }
     //Cannot have both conditionAnd and conditionOr properties
     if (cond['conditionAnd'] != undefined && cond['conditionOr'] != undefined) {
-      this.loggit.loggit('Condition has both AND and OR sub conditions. Not OK');
+      this.loggit.logLine('Condition has both AND and OR sub conditions. Not OK');
       isValid = false;
     }
     //Make sure Operator is valid
     if (!validOperators.includes(cond['operator'])) {
-      this.loggit.loggit('Condition Operator is not valid: ' + cond['operator']);
+      this.loggit.logLine('Condition Operator is not valid: ' + cond['operator']);
       isValid = false;
     }
     //Make sure Operators that require a comparitor have one
     if (operatorsNeedOperand.includes(cond['operator']) && isNaN(cond['operand'])) {
-      this.loggit.loggit('Operator ' + cond['operator'] + ' requires a value to operate against. Operand:' + JSON.stringify(cond['operand']));
+      this.loggit.logLine('Operator ' + cond['operator'] + ' requires a value to operate against. Operand:' + JSON.stringify(cond['operand']));
       isValid = false;
     }
 
@@ -210,14 +246,14 @@ export class packageInventory {
         cond.operand.sort();
       }
       else {
-        this.loggit.loggit('Operator ' + cond['operator'] + ' requires 2 operands');
+        this.loggit.logLine('Operator ' + cond['operator'] + ' requires 2 operands');
         isValid = false;
       }
     }
 
     //We need a metadata type unless operator is always or never
     if ((cond.operator != 'always' && cond.operator != 'never') && cond.metadataType == undefined) {
-      this.loggit.loggit('Operator ' + cond['operator'] + ' requires a metadataType');
+      this.loggit.logLine('Operator ' + cond['operator'] + ' requires a metadataType');
       isValid = false;
     }
     return isValid;
@@ -232,14 +268,14 @@ export class packageInventory {
   }
 
   public getDependencies() {
-    this.loggit.loggit('Checking Feature and License Dependencies');
+    this.loggit.logLine('Checking Feature and License Dependencies');
     if (!this._dependencies) {
       this._dependencies = [];
      
-      this.loggit.loggit('building dependencyrules from data models');
+      this.loggit.logLine('building dependencyrules from data models');
       let dependencyRulesConstructed = [...dependencyRules];
       for (var dataModel of dataModels) {
-        this.loggit.loggit(`Generating dependency rules from ${dataModel.name} data model`)
+        this.loggit.logLine(`Generating dependency rules from ${dataModel.name} data model`)
         
         let conditions = [];
         //Check Namespaces
@@ -267,7 +303,7 @@ export class packageInventory {
             operator: 'exists'
           });
         }
-        this.loggit.loggit(JSON.stringify(conditions));
+        this.loggit.logLine(JSON.stringify(conditions));
         if (conditions.length > 0) {
           let conditionConstructed = conditions.pop();
           let nextCondition;
@@ -281,12 +317,12 @@ export class packageInventory {
             label: dataModel.label,
             condition: conditionConstructed
           };
-          this.loggit.loggit('Generated Dependency Rule:');
-          this.loggit.loggit(JSON.stringify(dependencyConstructed));
+    //      this.loggit.logLine('Generated Dependency Rule:');
+    //      this.loggit.logLine(JSON.stringify(dependencyConstructed));
           dependencyRulesConstructed.push(dependencyConstructed) 
         }
       }
-      this.loggit.loggit('Adding Dependencies from Rules');
+      this.loggit.logLine('Adding Dependencies from Rules');
       for (var dependencyRule of dependencyRulesConstructed) {
         if (this.isConditionValid(dependencyRule.condition)) {
           if (this.checkCondition(dependencyRule.condition).conditionPass) {
@@ -298,7 +334,7 @@ export class packageInventory {
         }
       }
 
-     this.loggit.loggit('Adding Namespace Dependencies');
+     this.loggit.logLine('Adding Namespace Dependencies');
      let nameSpaces = this.getNamespaceReferences();
      if (nameSpaces.size > 0) {
        this._dependencies.push({
@@ -307,12 +343,7 @@ export class packageInventory {
          items: Array.from(nameSpaces)
        });
      }
-
-
     }
-
-    
-
     return this._dependencies;
   }
 
@@ -320,13 +351,13 @@ export class packageInventory {
     let nameSpaces = new Set();
     let nameSpaceResults = this.checkCondition({metadataType:'dependencies.namespaces.*',operator:'exists'});
     for (var ns of nameSpaceResults.passItems) {
-      nameSpaces.add(ns);
+      nameSpaces.add(ns.name);
     }
     //remove standard namespaces
-    nameSpaces.delete('c');
-    nameSpaces.delete('aura');
-    nameSpaces.delete('lightning');
-    nameSpaces.delete('ui');
+    for (const ns of standardNamespaces) {
+      nameSpaces.delete(ns);
+    }
+  
     return nameSpaces;
   }
 
@@ -334,11 +365,11 @@ export class packageInventory {
   public processRules(ruleSet) {
     //Replaces   public checkRules(ruleSet) 
     let results = [];
-    this.loggit.loggit('Checking Rules');
+    this.loggit.logLine('Checking Rules');
     for (var rule of ruleSet) {
-      this.loggit.loggit('Checking rule: ' + rule.name);
+      this.loggit.logLine('Checking rule: ' + rule.name);
       if (!this.isRuleValid(rule)) {
-        this.loggit.loggit('Rule is invalid. Skipping it.');
+        this.loggit.logLine('Rule is invalid. Skipping it.');
         continue;
       }
       let conditionResult = this.checkCondition(rule.condition);
@@ -351,7 +382,13 @@ export class packageInventory {
           result['url'] = rule.resultTrue.url;
         }
         if (rule.resultTrue.showDetails) {
-          result['exceptions'] = conditionResult.passItems;
+         // result['exceptions'] = conditionResult.passItems;
+         result['exceptions'] = [];
+         for (const i of conditionResult.passItems) {
+           if (i.display) {
+             result['exceptions'].push(i.name);
+           }
+         }
         }
         results.push(result);
       }
@@ -364,7 +401,13 @@ export class packageInventory {
           result['url'] = rule.resultFalse.url;
         }
         if (rule.resultFalse.showDetails) {
-          result['exceptions'] = conditionResult.failItems;
+          result['exceptions'] = [];
+       //   result['exceptions'] = conditionResult.failItems;
+          for (const i of conditionResult.failItems) {
+            if (i.display) {
+              result['exceptions'].push(i.name);
+            }
+          }
         }
         results.push(result);
       }
@@ -373,14 +416,14 @@ export class packageInventory {
   };
 
   public getInstallationWarnings() {
-    this.loggit.loggit('Getting Installation Warnings New Ruleset');
+    this.loggit.logLine('Getting Installation Warnings New Ruleset');
     if (!this._installationWarnings) {
       this._installationWarnings = [];
       for (var edition of editionWarningRules) {
-        this.loggit.loggit('Checking Edition rule: ' + JSON.stringify(edition));
+        this.loggit.logLine('Checking Edition rule: ' + JSON.stringify(edition));
         let editionBlock = [];
         for (var blockingItem of edition.blockingItems) {
-          this.loggit.loggit('Checking item:' + JSON.stringify(blockingItem));
+          this.loggit.logLine('Checking item:' + JSON.stringify(blockingItem));
           let result = this.checkCondition(blockingItem.condition);
           if (result.conditionPass) {
             editionBlock.push({
@@ -402,9 +445,9 @@ export class packageInventory {
   };
 
   public getTechAdoptionScore() {
-    this.loggit.loggit('Checking Tech Adoption Score');
+    this.loggit.logLine('Checking Tech Adoption Score');
     if (!this._techAdoption) {
-      this.loggit.loggit('Results not Cached');
+      this.loggit.logLine('Results not Cached');
     
     this._techAdoption = [...techAdoptionRules];
     for (var adoptionCategory of this._techAdoption) {
@@ -423,44 +466,44 @@ export class packageInventory {
   }
 
   public getAlerts() {
-    this.loggit.loggit('Checking Partner Alerts');
+    this.loggit.logLine('Checking Partner Alerts');
     if (!this._alerts) {
-      this.loggit.loggit('Results not Cached. Using New Rules Engine');
+      this.loggit.logLine('Results not Cached. Using New Rules Engine');
       this._alerts = this.processRules(alertRules);
     }
     return this._alerts;
   }
 
   public getQualityRecommendations() {
-    this.loggit.loggit('Checking Quality Recommendation Rules');
+    this.loggit.logLine('Checking Quality Recommendation Rules');
     //   return this.checkRules(qualityRules);
     if (!this._qualityRules) {
-      this.loggit.loggit('Results not Cached. Using New Rules Engine');
+      this.loggit.logLine('Results not Cached. Using New Rules Engine');
       this._qualityRules = this.processRules(qualityRules);
     }
     return this._qualityRules;
   };
 
   public getEnablementMessages() {
-    this.loggit.loggit('Checking Enablement Content Rules');
+    this.loggit.logLine('Checking Enablement Content Rules');
     //  return this.checkRules(enablementRules)
     if (!this._enablementMessages) {
-      this.loggit.loggit('Results not Cached. Using New Rules Engine');
+      this.loggit.logLine('Results not Cached. Using New Rules Engine');
       this._enablementMessages = this.processRules(enablementRules);
     }
     return this._enablementMessages;
   };
 
   public getCountByMetadataType(metadataType) {
-    this.loggit.loggit('getCountByMetadataType - Getting Count of Metadata by type: ' + metadataType);
+    this.loggit.logLine('getCountByMetadataType - Getting Count of Metadata by type: ' + metadataType);
     let mdDefArray = metadataType.split('.');
     let retVal = [];
     let mdCount = this.traverseMetadata(mdDefArray, this._mdInv);
     if (Array.isArray(mdCount)) {
-      this.loggit.loggit('Found an Array of results. We must have passed a wildcard search');
+      this.loggit.logLine('Found an Array of results. We must have passed a wildcard search');
       retVal = mdCount;
     } else {
-      this.loggit.loggit('Found a single result');
+      this.loggit.logLine('Found a single result');
       retVal.push(mdCount);
     }
     return retVal;
@@ -476,66 +519,66 @@ export class packageInventory {
     // if value exists and there are no more entries in the array, check to see if the value is a number
     // If it is a number, then return the number and the property name (or the wildcard name)
     // If it is not a number, then check to see if adding .count to the property is a number
-    this.loggit.loggit('Traversing Metadata');
-    this.loggit.loggit('Properties to traverse: ' + JSON.stringify(mdArray));
+    this.loggit.logLine('Traversing Metadata');
+    this.loggit.logLine('Properties to traverse: ' + JSON.stringify(mdArray));
     let topLevel = mdArray.shift();
-    this.loggit.loggit('Looking at: ' + topLevel);
+    this.loggit.logLine('Looking at: ' + topLevel);
     if (topLevel === '*') {
       let retVal = [];
-      this.loggit.loggit('Checking Wildcard');
+      this.loggit.logLine('Checking Wildcard');
 
       for (var topArray in mdObject) {
         let tmpArray = [topArray, ...mdArray];
 
-        this.loggit.loggit('checking new Array: ' + tmpArray);
-        this.loggit.loggit(`ParamArray: ${tmpArray}, Wildcard: ${topArray}`);
+        this.loggit.logLine('checking new Array: ' + tmpArray);
+        this.loggit.logLine(`ParamArray: ${tmpArray}, Wildcard: ${topArray}`);
         retVal.push(this.traverseMetadata(tmpArray, mdObject, topArray));
       }
-      this.loggit.loggit('Wildcard Processed');
+      this.loggit.logLine('Wildcard Processed');
       return retVal;
     } else if (mdObject[topLevel] != undefined) {
       if (mdArray.length > 0) {
-        this.loggit.loggit('There are more components left.');
+        this.loggit.logLine('There are more components left.');
         if (mdArray.length == 1 && mdArray[0] === '*') {
-          this.loggit.loggit('Last component is a wildcard. Peeking ahead');
+          this.loggit.logLine('Last component is a wildcard. Peeking ahead');
           //Check to see if the wildcard object has any values
           if (Object.keys(mdObject[topLevel]).length == 0) {
-            this.loggit.loggit(` Component ${topLevel} is an empty object. We can stop here`);
+            this.loggit.logLine(` Component ${topLevel} is an empty object. We can stop here`);
             return {
               property: topLevel,
               value: -1
             };
           }
         }
-        this.loggit.loggit('We can safely Recurse');
-        this.loggit.loggit(`  ObjectKey: ${topLevel} ParamArray: ${mdArray}, Wildcard: ${wildcard}`);
+        this.loggit.logLine('We can safely Recurse');
+        this.loggit.logLine(`  ObjectKey: ${topLevel} ParamArray: ${mdArray}, Wildcard: ${wildcard}`);
         return this.traverseMetadata(mdArray, mdObject[topLevel], wildcard);
       } else {
-        this.loggit.loggit('This is last portion. Looking for value');
+        this.loggit.logLine('This is last portion. Looking for value');
         let count = undefined;
         if (isNaN(mdObject[topLevel])) {
-          this.loggit.loggit(`${JSON.stringify(mdObject[topLevel])} is not a number.`);
+          this.loggit.logLine(`${JSON.stringify(mdObject[topLevel])} is not a number.`);
           if (mdObject[topLevel]['count'] != undefined && isFinite(mdObject[topLevel]['count'])) {
-            this.loggit.loggit('')
+   //         this.loggit.logLine('')
             count = mdObject[topLevel]['count'];
           } else if (mdObject[topLevel] === Object(mdObject[topLevel])) {
             //if it's an object, return the number of keys
-            this.loggit.loggit(`${topLevel} appears to be an object. Returning count of keys`);
+            this.loggit.logLine(`${topLevel} appears to be an object. Returning count of keys`);
             count = Object.keys(mdObject[topLevel]).length;
           } else if (Array.isArray(mdObject[topLevel])) {
-            this.loggit.loggit(`${topLevel} appears to be an array. Returning array count`);
+            this.loggit.logLine(`${topLevel} appears to be an array. Returning array count`);
             count = mdObject[topLevel].length;
           }
           else {
-            this.loggit.loggit(' Cannot find a valid number returning empty object');
+            this.loggit.logLine(' Cannot find a valid number returning empty object');
             return {};
           }
 
         } else {
-          this.loggit.loggit('Using value from ' + topLevel);
+          this.loggit.logLine('Using value from ' + topLevel);
           count = mdObject[topLevel];
         }
-        this.loggit.loggit('Final Value:' + count);
+        this.loggit.logLine('Final Value:' + count);
         let componentName = wildcard == '' ? topLevel : wildcard;
         return {
           property: componentName,
@@ -544,7 +587,7 @@ export class packageInventory {
       }
 
     } else {
-      this.loggit.loggit(`could not find Key ${topLevel} in the object.`);
+      this.loggit.logLine(`could not find Key ${topLevel} in the object.`);
       return {
         property: topLevel,
         value: -1
@@ -552,9 +595,9 @@ export class packageInventory {
     }
   };
 
-
+/*
   public getInventoryCountByMetadataType(metadataType) {
-    this.loggit.loggit('getInventoryCountByMetadataType - Getting Count of mdType: ' + metadataType);
+    this.loggit.logLine('getInventoryCountByMetadataType - Getting Count of mdType: ' + metadataType);
     let mdType = this.parseMetadataType(metadataType);
     let count = 0;
     if (this._mdInv[mdType.metadataType] != undefined) {
@@ -573,7 +616,7 @@ export class packageInventory {
   };
 
   public parseMetadataType(metadataType) {
-    this.loggit.loggit('Parsing Metadata Type' + metadataType);
+    this.loggit.logLine('Parsing Metadata Type' + metadataType);
     let mdType = metadataType.split('.'); //To look at subtypes, use Type.Subtype (e.g.: ApexClass.BatchApex)
     let key = mdType[1] ? mdType[1] : 'count';
     let parsedType = {
@@ -586,12 +629,12 @@ export class packageInventory {
     if (mdType[0] === 'apiVersions' && mdType[2]) {
       parsedType['targetComponent'] = mdType[2];
     }
-    this.loggit.loggit('Parsed Result: ' + JSON.stringify(parsedType));
+    this.loggit.logLine('Parsed Result: ' + JSON.stringify(parsedType));
     return parsedType;
   };
-
+*/
   public getJSONOutput() {
-    this.loggit.loggit('Getting JSON output');
+    this.loggit.logLine('Getting JSON output');
     let retVal = {};
     retVal['MetadataInventory'] = this._mdInv;
     retVal['MonitoredItems'] = this.getMonitoredInvArray();
@@ -606,7 +649,7 @@ export class packageInventory {
 
 
   public getFullInvArray() {
-    this.loggit.loggit('Getting Full Inventory');
+    this.loggit.logLine('Getting Full Inventory');
     if (!this._mdFullInvArray) {
       this._mdFullInvArray = [];
       for (var mdType in this._mdInv) {
@@ -623,11 +666,11 @@ export class packageInventory {
   };
 
   public getMonitoredInvArray() {
-    this.loggit.loggit('Getting Monitored Inventory');
+    this.loggit.logLine('Getting Monitored Inventory');
     if (!this._mdMonitoredInvArray) {
       this._mdMonitoredInvArray = [];
       mdTypes.forEach(element => {
-        this.loggit.loggit('Processing Monitored Item: ' + JSON.stringify(element));
+        this.loggit.logLine('Processing Monitored Item: ' + JSON.stringify(element));
         let retObj = {};
         let extras = [];
         let extrasCustom = [];
@@ -639,7 +682,7 @@ export class packageInventory {
           switch (String(element.metadataType)) {
 
             case 'CustomField':
-              this.loggit.loggit('Checking Custom Fields');
+              this.loggit.logLine('Checking Custom Fields');
               if (this._mdInv[element.metadataType]) {
                 count = this._mdInv[element.metadataType]['count'];
                 const objects = Object.keys(this._mdInv[element.metadataType]['objects']);
