+++ src/common/metadataScan.ts
@@ -1,0 +1,1082 @@
+
+import {
+  SfdxError
+} from '@salesforce/core';
+
+import {
+  setValue,
+  incrementValue,
+  getValue
+} from './JSONUtilities';
+
+import {
+  Loggit
+} from './logger';
+
+import {
+  mdmap
+} from './mdmap';
+
+import fs = require('fs-extra');
+
+import xml2js = require('xml2js');
+
+
+export function inventoryPackage(sourceDir, p) {
+  let types = p.types;
+  let inventory = {};
+  let apiVersions = {};
+  let componentProperties = {};
+  let dependencies = {};
+  let loggit = new Loggit('isvtePlugin:metadataScan');
+
+  if (p.version) {
+    apiVersions['mdapi'] = parseFloat(p.version[0]);
+  }
+  for (var typeIdx in types) {
+    let metadataType = types[typeIdx]['name'];
+    
+    let typeInv = {};
+
+    typeInv['index'] = typeIdx;
+
+    loggit.logLine('Checking MetadataType: ' + metadataType);
+  
+    //Check for wildcard members
+    if (types[typeIdx]['members'].includes('*')) {
+      loggit.logLine('Found Wildcard Members');
+      types[typeIdx]['members'] = getMembers(types[typeIdx], sourceDir);
+      //        loggit.loggit('Members: ' + JSON.stringify(types[typeIdx]['members']));
+    }
+    typeInv['count'] = types[typeIdx]['members'].length;
+
+    loggit.logLine('  Found ' + types[typeIdx]['members'].length + ' members');
+    loggit.logLine('Members: ' + JSON.stringify(types[typeIdx]['members']));
+
+    switch (String(metadataType)) {
+      case 'CustomField':
+        //Do per object field counts
+        let objectFields = {};
+
+        for (var fieldIdx in types[typeIdx]['members']) {
+          let fieldFullName = types[typeIdx]['members'][fieldIdx];
+          let object = getNameSpaceAndType(fieldFullName.split(".")[0]);
+          let field = getNameSpaceAndType(fieldFullName.split(".")[1]);
+          addObjectDependencies(dependencies,[object,field]);
+/*
+          //Add Namespace Dependencies 
+          if (object.namespace !== null) {
+            if (dependencies['namespaces'] == undefined) {
+              dependencies['namespaces'] = {};
+            }
+            dependencies['namespaces'][object.namespace] = 1;
+          }
+          if (field.namespace !== null) {
+            if (dependencies['namespaces'] == undefined) {
+              dependencies['namespaces'] = {};
+            }
+            dependencies['namespaces'][field.namespace] = 1;
+          }
+
+          if (dependencies['components'] == undefined) {
+            dependencies['components'] = {};
+          }
+          dependencies['components'][object.fullName] = object;
+
+          if (dependencies['components'] == undefined) {
+            dependencies['components'] = {};
+          }
+          dependencies['components'][field.fullName] = field;
+         */
+          if (objectFields[object.fullName]) {
+            objectFields[object.fullName]['count'] += 1;
+          } else {
+            objectFields[object.fullName] = {
+              'count': 1,
+              'objectType': object.type
+            };
+          }
+
+          //Check field descriptions
+          //Only check custom fields or standard fields on custom objects, not standard
+          //Never check for namespaced fields
+          if (field.namespace == null && ((object.type == 'Standard' && field.type !== 'Standard')|| object.type !== 'Standard')) {
+            const objectPath = `${sourceDir}/objects`;
+            let objectXml = `${objectPath}/${object.fullName}.object`;
+            let objectJSON = parseXML(objectXml);
+            if (objectJSON['CustomObject'] && objectJSON['CustomObject']['fields']) {
+              for (var fieldDef of objectJSON['CustomObject']['fields']) {
+                if (fieldDef['fullName'] == field.fullName) {
+                  loggit.logLine('Checking Properties of Field: ' + fieldFullName);
+/*
+                  if (componentProperties['CustomField'] == undefined) {
+                    componentProperties['CustomField'] = {};
+                  }
+                  if (componentProperties['CustomField'][fieldFullName] == undefined) {
+                    componentProperties['CustomField'][fieldFullName] = {};
+                  }
+                  componentProperties['CustomField'][fieldFullName]['descriptionExists'] = fieldDef['description'] ? 1 : 0;*/
+                  let descExists = fieldDef['description'] ? 1 : 0;
+                  setValue(componentProperties,`CustomField.${fieldFullName.replace('.','->')}.descriptionExists`,descExists);
+                }
+
+              }
+            }
+          }
+        }
+        typeInv['objects'] = objectFields;
+
+        break;
+      case 'CustomObject':
+        // Look for Custom Settings, External Objects,  Person Accounts, Big Objects
+        loggit.logLine('Deep Inventory on Custom Objects');
+        const objectPath = `${sourceDir}/objects`;
+        //let xoType = {count:0};
+        let xoCount = 0;
+        //let boType = {count:0};
+        let boCount = 0;
+        let csType = {
+          count: 0
+        };
+        let peType = {
+          count: 0
+        };
+        let fmType = {
+          count: 0
+        };
+        for (var objIdx in types[typeIdx]['members']) {
+          let object = getNameSpaceAndType(types[typeIdx]['members'][objIdx]);
+          loggit.logLine('Checking Object: ' + object.fullName);
+          addObjectDependencies(dependencies,[object]);
+/*          //Add Namespace Dependencies 
+          if (object.namespace !== null) {
+            if (dependencies['namespaces'] == undefined) {
+              dependencies['namespaces'] = {};
+            }
+            dependencies['namespaces'][object.namespace] = 1;
+          }
+
+          if (dependencies['components'] == undefined) {
+            dependencies['components'] = {};
+          }
+          dependencies['components'][object.fullName] = object;
+*/
+          //Check external Objects
+          if (object.extension == 'e') {
+            xoCount += 1;
+          }
+          //Check Big Objects
+          if (object.extension == 'Big Object') {
+            boCount += 1;
+          }
+
+          //Check Platform Events
+          if (object.extension == 'e') {
+            peType['count'] += 1;
+          }
+
+          //Check Feature Management Parameters
+          if (String(object.fullName).includes('FeatureParameter')) {
+            fmType['count'] += 1;
+          }
+
+          let objectXml = `${objectPath}/${object.fullName}.object`;
+          let objectJSON = parseXML(objectXml);
+
+          //Check Custom Settings
+          if (objectJSON['CustomObject'] && objectJSON['CustomObject']['customSettingsType']) {
+            csType['count'] + 1;
+          }
+          //Check for Descriptions
+          if (object.type == 'Custom' && object.namespace == null) {
+            loggit.logLine('Checking properties of object ' + object.fullName);
+/*
+            if (componentProperties['CustomObject'] == undefined) {
+              componentProperties['CustomObject'] = {};
+            }
+            if (componentProperties['CustomObject'][object.fullName] == undefined) {
+              componentProperties['CustomObject'][object.fullName] = {};
+            }
+            componentProperties['CustomObject'][object.fullName]['descriptionExists'] = objectJSON['CustomObject'] && objectJSON['CustomObject']['description'] ? 1 : 0;*/
+            let descExists = getValue(objectJSON,'CustomObject.description',null) ? 1 : 0;
+            setValue(componentProperties,`CustomObject.${object.fullName}.descriptionExists`,descExists);
+
+          }
+
+          //   loggit(objectJSON,'JSON');
+        }
+        //inventory['ExternalObject__c'] = xoType;
+        //inventory['BigObject__c'] = boType;
+
+        typeInv['BigObject'] = boCount;
+        typeInv['ExternalObject'] = xoCount;
+        inventory['CustomSetting__c'] = csType;
+        inventory['PlatformEvent__c'] = peType;
+        inventory['FeatureManagement__c'] = fmType;
+
+        break;
+      case 'PlatformEventChannel':
+        //Look for ChangeDataEvents
+        break;
+      case 'Flow':
+        //Check for Flow Templates
+
+        loggit.logLine('Checking flows');
+
+        let templateCount = 0;
+     //   let screenTemplateCount = 0;
+     //   let autolaunchedTemplateCount = 0;
+        let objects = {};
+        let flowTypes = {};
+        let flowTemplates = {};
+             const flowPath = `${sourceDir}/flows`;
+        for (var flowIdx in types[typeIdx]['members']) {
+          let flowName = types[typeIdx]['members'][flowIdx];
+          let flowXml = `${flowPath}/${flowName}.flow`;
+          let flowJSON = parseXML(flowXml);
+          loggit.logLine('Checking file:' + flowXml);
+       //   let processType = flowJSON['Flow'] && flowJSON['Flow']['processType'] ? flowJSON['Flow']['processType'] : 'UnknownType';
+          let processType = getValue(flowJSON,'Flow.processType','UnknownType');
+          incrementValue(flowTypes,`${processType}.count`);
+       
+          
+          if (flowJSON['Flow'] && flowJSON['Flow']['isTemplate'] && flowJSON['Flow']['isTemplate'][0] === 'true') {
+            templateCount += 1;
+            incrementValue(flowTemplates,`${processType}.count`);
+           /* if (flowTemplates[processType]) {
+              flowTemplates[processType]['count']+=1;
+            }
+            else {
+              flowTemplates[processType] = {
+                count:1
+              };
+            }*/
+          }
+
+          //Do per object Inventory
+          loggit.logLine('Inventorying PB and Flow Triggers Per Object');
+          loggit.logLine('Flow Details: ' + JSON.stringify(flowJSON['Flow']['processMetadataValues']));
+            if (flowJSON['Flow']['processMetadataValues']) {
+              for (var processMetadataValue of flowJSON['Flow']['processMetadataValues']) {
+                loggit.logLine('Metadata Value Name: ' + processMetadataValue['name']);
+                if (processMetadataValue['name'] == 'ObjectType') {
+                  loggit.logLine('ObjectName:' + JSON.stringify(processMetadataValue['value'][0]));
+                  let objectName = processMetadataValue['value'][0]['stringValue'][0];
+                  let object = getNameSpaceAndType(objectName);
+                  addObjectDependencies(dependencies,[object]);
+                  //Add Namespace Dependencies 
+               /*   if (object.namespace !== null) {
+                    if (dependencies['namespaces'] == undefined) {
+                      dependencies['namespaces'] = {};
+                    }
+                    dependencies['namespaces'][object.namespace] = 1;
+                  }
+                  if (dependencies['components'] == undefined) {
+                    dependencies['components'] = {};
+                  }
+                  dependencies['components'][object.fullName] = object;
+*/
+                  loggit.logLine('Extracted Object Name:' + objectName);
+                  incrementValue(objects,`${objectName}.count`);
+               /*   if (objects[objectName]) {
+                    objects[objectName]['count'] += 1;
+                  } else {
+                    objects[objectName] = {
+                      count: 1
+                    };
+                  }*/
+                }
+              }
+            }
+            
+          
+        }
+        typeInv['FlowTypes'] = flowTypes;
+        typeInv['FlowTemplates'] = flowTemplates;
+        typeInv['FlowTemplate'] = templateCount;
+        typeInv['objects'] = objects;
+        break;
+      case 'CustomApplication':
+        let lightningCount = 0;
+        let classicCount = 0;
+        let lightingConsoleCount = 0;
+        let classicConsoleCount = 0;
+        for (var appIdx in types[typeIdx]['members']) {
+          let appName = types[typeIdx]['members'][appIdx];
+          let uiType;
+          let navType;
+          loggit.logLine('Checking App: ' + appName);
+          const appPath = `${sourceDir}/applications`;
+          let appXml = `${appPath}/${appName}.app`;
+          let appJSON = parseXML(appXml);
+          if (appJSON['CustomApplication']) {
+            if (appJSON['CustomApplication']['uiType']) {
+              uiType = appJSON['CustomApplication']['uiType'][0];
+            }
+            if (appJSON['CustomApplication']['navType']) {
+              navType = appJSON['CustomApplication']['navType'][0];
+            }
+            if (uiType === 'Lightning') {
+              lightningCount += 1;
+              if (navType === 'Console') {
+                lightingConsoleCount += 1;
+              }
+            } else if (uiType === 'Aloha') {
+              classicCount += 1;
+              if (navType === 'Console') {
+                classicConsoleCount += 1;
+              }
+            }
+          }
+        }
+        typeInv['LightingAppCount'] = lightningCount;
+        typeInv['LightningConsoleCount'] = lightingConsoleCount;
+        typeInv['ClassicAppCount'] = classicCount;
+        typeInv['ClassicConsoleCount'] = classicConsoleCount;
+        break;
+      case 'ConnectedApp':
+        loggit.logLine('Checking Connected Apps');
+
+        let canvasCount = 0;
+
+        const caPath = `${sourceDir}/connectedApps`;
+        for (var caIdx in types[typeIdx]['members']) {
+          let caName = types[typeIdx]['members'][caIdx];
+          let caXml = `${caPath}/${caName}.connectedApp`;
+          let caJSON = parseXML(caXml);
+          if (caJSON['ConnectedApp'] && caJSON['ConnectedApp']['canvasConfig']) {
+            canvasCount += 1;
+          }
+        }
+        typeInv['CanvasApp'] = canvasCount;
+        break;
+      case 'ApexClass':
+        loggit.logLine('Interrogating Apex');
+        let futureCount = 0;
+        // let testCount = 0;
+        let auraEnabledCount = 0;
+        let batchCount = 0;
+        let schedulableCount = 0;
+        let invocableCount = 0;
+        let apexRestCount = 0;
+        let apexSoapCount = 0;
+
+
+        const apexPath = `${sourceDir}/classes`;
+        for (var apxIdx in types[typeIdx]['members']) {
+          let className = types[typeIdx]['members'][apxIdx];
+          let classFile = `${apexPath}/${className}.cls`;
+          if (fs.existsSync(classFile)) {
+
+            let classBody = stripApexComments(fs.readFileSync(classFile, 'utf8'));
+            // loggit(classBody);
+            //const testReg = /@istest/ig;
+            const futureReg = /@future/ig;
+            const auraEnabledReg = /@AuraEnabled/ig;
+            const invocableReg = /@InvocableMethod|InvocableVariable/ig;
+            const batchReg = /implements\s+Database\.Batchable/ig;
+            const scheduleReg = /implements\s+Schedulable/ig;
+            const restReg = /@RestResource/ig;
+            const soapReg = /webservice\s+static/ig;
+            const advFLSSOQLReg = /SECURITY_ENFORCED/ig;
+            const advFLSStripInaccessible = /Security\.stripInaccessible/ig;
+          //  const refersGuestSimpleReg = /UserType(?:\(\))?\s*=\s*(['"])Guest\1/ig;
+          //  const refersGuestComplexReg = /(\w+)\s*=.*getUserType\(\)(?:.*)\1\s*=\s*(["'])Guest\2/is;
+            const refersGuestTrivialReg = /(["'])Guest\1/ig;
+           
+            
+            if (futureReg.test(classBody)) {
+              futureCount += 1;
+            }
+            if (auraEnabledReg.test(classBody)) {
+              auraEnabledCount += 1;
+            }
+            if (invocableReg.test(classBody)) {
+              invocableCount += 1;
+            }
+            if (restReg.test(classBody)) {
+              apexRestCount += 1;
+            }
+            if (soapReg.test(classBody)) {
+              apexSoapCount += 1;
+            }
+            if (scheduleReg.test(classBody)) {
+              schedulableCount += 1;
+            }
+            if (batchReg.test(classBody)) {
+              batchCount += 1;
+            }
+            if (advFLSSOQLReg.test(classBody)) {
+              setValue(componentProperties,`ApexClass.${className}.SECURITY_ENFORCED`,1);
+              /* if (componentProperties['ApexClass'] == undefined) {
+                componentProperties['ApexClass'] = {};
+              }
+              if (componentProperties['ApexClass'][className] == undefined) {
+                componentProperties['ApexClass'][className] = {};
+              }
+              componentProperties['ApexClass'][className]['SECURITY_ENFORCED'] = 1;*/
+            }
+            if (advFLSStripInaccessible.test(classBody)) {
+              setValue(componentProperties,`ApexClass.${className}.StripInaccessible`,1);
+/*
+              if (componentProperties['ApexClass'] == undefined) {
+                componentProperties['ApexClass'] = {};
+              }
+              if (componentProperties['ApexClass'][className] == undefined) {
+                componentProperties['ApexClass'][className] = {};
+              }
+              componentProperties['ApexClass'][className]['StripInaccessible'] = 1;*/
+            }
+
+        //    if (refersGuestComplexReg.test(classBody) || refersGuestSimpleReg.test(classBody)) {
+            if (refersGuestTrivialReg.test(classBody)) {
+              setValue(componentProperties,`ApexClass.${className}.RefersToGuest`,1);
+/*
+              if (componentProperties['ApexClass'] == undefined) {
+                componentProperties['ApexClass'] = {};
+              }
+              if (componentProperties['ApexClass'][className] == undefined) {
+                componentProperties['ApexClass'][className] = {};
+              }*/
+            }
+            //Find Object References
+            loggit.logLine('Looking for fields and objects referenced within APEX')
+            
+            addObjectDependencies(dependencies, extractObjectsApex(classBody));
+                    
+          }
+
+          let classMetaFile = `${apexPath}/${className}.cls-meta.xml`;
+          if (fs.existsSync(classMetaFile)) {
+            let classMetaJSON = parseXML(classMetaFile);
+            if (classMetaJSON['ApexClass'] && classMetaJSON['ApexClass']['apiVersion']) {
+        /*      if (apiVersions['ApexClass'] == undefined) {
+                apiVersions['ApexClass'] = {};
+              }*/
+              setValue(apiVersions,`ApexClass.${className}`,parseFloat(classMetaJSON['ApexClass']['apiVersion'][0]));
+         //     addKeyToObject(apiVersions,'ApexClass');
+         //     apiVersions['ApexClass'][className] = parseFloat(classMetaJSON['ApexClass']['apiVersion'][0]);
+            }
+          }
+        }
+        typeInv['FutureCalls'] = futureCount;
+        typeInv['AuraEnabledCalls'] = auraEnabledCount;
+        typeInv['InvocableCalls'] = invocableCount;
+        // typeInv['TestMethods'] = testCount;
+        typeInv['BatchApex'] = batchCount;
+        typeInv['SchedulableApex'] = schedulableCount;
+        typeInv['ApexRest'] = apexRestCount;
+        typeInv['ApexSoap'] = apexSoapCount;
+
+
+        break;
+      case 'ApexTrigger':
+        loggit.logLine('Interrogating Trigger');
+        let triggerInv = {};
+        //let asyncTrigger = {'count':0};
+        let asyncCount = 0;
+        const triggerPath = `${sourceDir}/triggers`;
+        for (var triggerIdx in types[typeIdx]['members']) {
+          let triggerName = types[typeIdx]['members'][triggerIdx];
+          let triggerFile = `${triggerPath}/${triggerName}.trigger`;
+          let triggerBody = stripApexComments(fs.readFileSync(triggerFile, 'utf8'));
+          const triggerDetailReg = /trigger\s+(\w+)\s+on\s+(\w+)\s*\((.+)\)/im;
+          const refersGuestTrivialReg = /(["'])Guest\1/ig;
+          const findObjectsReg = /(?:(?<namespace>[a-zA-Z](?:[a-z]|[A-Z]|[0-9]|_(?!_)){0,14})__)?(?<component>(?<!___)[a-zA-Z](?:[a-z]|[A-Z]|[0-9]|_(?!_))+)(?:__(?<suffix>c|mdt|e|x|b|pc|pr|r|xo|latitude__s|longitude__s|history|ka|kav|feed|share))/g;
+
+
+          let triggerDetail = triggerDetailReg.exec(triggerBody);
+          if (triggerDetail == null) {
+            loggit.logLine('Could not parse Trigger File: ' + triggerFile);
+          }
+          else {
+            let triggerObject = getNameSpaceAndType(triggerDetail[2]);
+            let triggerType = triggerDetail[3];
+            loggit.logLine('Trigger Name:' + triggerName);
+            loggit.logLine('Trigger Object:' + triggerObject.fullName);
+            loggit.logLine('Trigger Type: ' + triggerType);
+            //Add Namespace Dependencies 
+       /*     if (triggerObject.namespace !== null) {
+              if (dependencies['namespaces'] == undefined) {
+              dependencies['namespaces'] = {};
+              }
+              dependencies['namespaces'][triggerObject.namespace] = 1;
+            }
+            if (dependencies['components'] == undefined) {
+              dependencies['components'] = {};
+            }
+            dependencies['components'][triggerObject.fullName] = triggerObject;
+*/
+            addObjectDependencies(dependencies,[triggerObject]);
+            if (triggerObject.type == 'Change Data Capture') {
+              //  asyncTrigger['count']++;
+              asyncCount += 1;
+            }
+            incrementValue(triggerInv,`${triggerObject.fullName}.count`);
+            /*
+           if (triggerInv[triggerObject.fullName]) {
+              triggerInv[triggerObject.fullName]['count'] += 1;
+            } else {
+              triggerInv[triggerObject.fullName] = {
+                count: 1
+              };
+            }*/
+            if (refersGuestTrivialReg.test(triggerBody)) {
+              setValue(componentProperties,`ApexTrigger.${triggerName}.refersToGuest`,1);
+         /*     if (componentProperties['ApexTrigger'] == undefined) {
+                componentProperties['ApexTrigger'] = {};
+              }
+              if (componentProperties['ApexTrigger'][triggerName] == undefined) {
+                componentProperties['ApexTrigger'][triggerName] = {};
+              }
+              componentProperties['ApexTrigger'][triggerName]['RefersToGuest'] = 1;*/
+            }
+
+            //Find Object References
+           // let objectFound;
+            loggit.logLine('Looking for fields and objects referenced within APEX')
+            
+            addObjectDependencies(dependencies, extractObjectsApex(triggerBody));
+/*
+            while (objectFound = findObjectsReg.exec(triggerBody)) {
+              loggit.logLine('Found: ' + objectFound[0]);
+              const object = getNameSpaceAndType(objectFound[0]);
+              //Add Namespace Dependencies 
+              if (object.namespace !== null) {
+                if (dependencies['namespaces'] == undefined) {
+                dependencies['namespaces'] = {};
+                }
+                dependencies['namespaces'][object.namespace] = 1;
+              }
+              if (dependencies['components'] == undefined) {
+                dependencies['components'] = {};
+              }
+              dependencies['components'][object.fullName] = object;
+            }        
+*/
+          }
+
+          let triggerMetaFile = `${triggerPath}/${triggerName}.trigger-meta.xml`;
+          if (fs.existsSync(triggerMetaFile)) {
+            let triggerMetaJSON = parseXML(triggerMetaFile);
+            if (triggerMetaJSON['ApexTrigger'] && triggerMetaJSON['ApexTrigger']['apiVersion']) {
+/*
+              if (apiVersions['ApexTrigger'] == undefined) {
+                apiVersions['ApexTrigger'] = {};
+              }*/
+              setValue(apiVersions,`ApexTrigger.${triggerName}`,parseFloat(triggerMetaJSON['ApexTrigger']['apiVersion'][0]));
+
+        //      apiVersions['ApexTrigger'][triggerName] = parseFloat(triggerMetaJSON['ApexTrigger']['apiVersion'][0]);
+            }
+          }
+        }
+        typeInv['objects'] = triggerInv;
+        typeInv['AsyncTrigger'] = asyncCount;
+        break;
+      case 'LightningComponentBundle':
+        loggit.logLine('Interrogating LWC');
+        const lwcPath = `${sourceDir}/lwc`;
+        let exposedCount = 0;
+        let targets = {}
+
+
+        for (var lwcIdx in types[typeIdx]['members']) {
+          const lwcName = types[typeIdx]['members'][lwcIdx];
+          const lwcXml = `${lwcPath}/${lwcName}/${lwcName}.js-meta.xml`;
+          //TODO: Parse for object and namespace references
+          let lwcJSON = parseXML(lwcXml);
+          if (lwcJSON['LightningComponentBundle']) {
+            lwcJSON = lwcJSON['LightningComponentBundle'];
+
+            loggit.logLine('Checking LWC ' + lwcName);
+            // loggit(lwcJSON,'JSON');
+            if (lwcJSON['apiVersion']) {
+       /*       if (apiVersions['LightningComponentBundle'] == undefined) {
+                apiVersions['LightningComponentBundle'] = {};
+              }
+              apiVersions['LightningComponentBundle'][lwcName] = parseFloat(lwcJSON['apiVersion'][0]);*/
+              setValue(apiVersions,`LightningComponentBundle.${lwcName}`,parseFloat(lwcJSON['apiVersion'][0]));
+
+            }
+            if (lwcJSON['isExposed'] && lwcJSON['isExposed'][0] === 'true') {
+              exposedCount += 1;
+            }
+            if (lwcJSON['targets'] && lwcJSON['targets'][0]['target']) {
+              loggit.logLine('Checking Targets');
+              loggit.logLine(lwcJSON['targets'][0]);
+              for (let target of lwcJSON['targets'][0]['target']) {
+                if (targets[target] != undefined) {
+                  targets[target] += 1;
+                }
+                else {
+                  targets[target] = 1;
+                }
+              }
+            }
+          }
+        }
+        typeInv['ExposedComponents'] = exposedCount;
+        typeInv['targets'] = targets;
+
+        break;
+      case 'ApexPage':
+        loggit.logLine('Interrogating Visualforce');
+        const vfPath = `${sourceDir}/pages`;
+        for (var vfIdx in types[typeIdx]['members']) {
+          const vfName =  types[typeIdx]['members'][vfIdx];
+          const vfFile = `${vfPath}/${vfName}.page`;
+          const vfXML = `${vfPath}/${vfName}.page-meta.xml`;
+          if (fs.existsSync(vfXML)) {
+            let vfMetaJSON = parseXML(vfXML);
+            if (vfMetaJSON['ApexPage'] && vfMetaJSON['ApexPage']['apiVersion']) {
+/*
+              if (apiVersions['ApexPage'] == undefined) {
+                apiVersions['ApexPage'] = {};
+              }
+              apiVersions['ApexPage'][vfName] = parseFloat(vfMetaJSON['ApexPage']['apiVersion'][0]);*/
+              setValue(apiVersions,`ApexPage.${vfName}`,parseFloat(vfMetaJSON['ApexPage']['apiVersion'][0]));
+
+            }
+          }
+          if (fs.existsSync(vfFile)) {
+            const vfBody = fs.readFileSync(vfFile, 'utf8');
+            const referSiteReg = /{!.*(\$Site|\$Network).*}/ig;
+            const stdControllerReg = /standardController="([a-zA-Z0-9_]+)"/i;
+            
+            //Find Standard Controllers
+            let stdControllerMatch = stdControllerReg.exec(vfBody);
+
+            if (stdControllerMatch !== null) {
+              let controllerObject = getNameSpaceAndType(stdControllerMatch[1]);
+              console.log('Found a standard Controller!!:' + JSON.stringify(stdControllerMatch));
+              addObjectDependencies(dependencies,[controllerObject]);
+        /*      //Add Namespace Dependencies 
+              if (controllerObject.namespace !== null) {
+                if (dependencies['namespaces'] == undefined) {
+                dependencies['namespaces'] = {};
+                }
+                dependencies['namespaces'][controllerObject.namespace] = 1;
+              }
+              if (dependencies['components'] == undefined) {
+                dependencies['components'] = {};
+              }
+              dependencies['components'][controllerObject.fullName] = controllerObject;*/
+            }
+
+            //Find namespaces in components used
+            loggit.logLine('Performing Regex search against component for namespaces');
+            const componentsReg = /<(\w+:\w+)/ig;
+            let referencedComponents = getMatches(vfBody, componentsReg);
+            if (referencedComponents.length > 0) {
+              loggit.logLine(`Found the following Components: ${JSON.stringify(referencedComponents)}`);
+           //   addKeyToObject(componentProperties,`ApexPage.${vfName}.namespaceReferences`);
+           /*   if (componentProperties['ApexPage'] == undefined) {
+                componentProperties['ApexPage'] = {};
+              }
+              if (componentProperties['ApexPage'][vfName] == undefined) {
+                componentProperties['ApexPage'][vfName] = {};
+              }
+              if (componentProperties['ApexPage'][vfName]['namespaceReferences'] == undefined) {
+                componentProperties['ApexPage'][vfName]['namespaceReferences'] = {};
+              }*/
+              referencedComponents.forEach(element => {
+                let ns = element.split(":", 2)[0];
+                incrementValue(componentProperties,`ApexPage.${vfName}.namespaceReferences.${ns}`);
+              /*  if (componentProperties['ApexPage'][vfName]['namespaceReferences'][ns] == undefined) {
+                  componentProperties['ApexPage'][vfName]['namespaceReferences'][ns] = 1;
+                }
+                else {
+                  componentProperties['ApexPage'][vfName]['namespaceReferences'][ns] += 1
+                }*/
+                //Also add it to the Namespaces dependencies
+                setValue(dependencies,`namespaces.${ns}`,1);
+           /*     if (dependencies['namespaces'] == undefined) {
+                  dependencies['namespaces'] = {};
+                }
+                dependencies['namespaces'][ns] = 1;*/
+              });
+            }
+
+            if (referSiteReg.test(vfBody)) {
+              setValue(componentProperties,`ApexPage.${vfName}.RefersToSite`,1);
+          /*    if (componentProperties['ApexPage'] == undefined) {
+                componentProperties['ApexPage'] = {};
+              }
+              if (componentProperties['ApexPage'][vfName] == undefined) {
+                componentProperties['ApexPage'][vfName] = {};
+              }
+              componentProperties['ApexPage'][vfName]['RefersToSite'] = 1;*/
+            }
+
+          }
+        }
+        break;
+      case 'AuraDefinitionBundle':
+        loggit.logLine('Interrogating Aura Components');
+        const auraPath = `${sourceDir}/aura`;
+        for (var auraIdx in types[typeIdx]['members']) {
+          const auraName = types[typeIdx]['members'][auraIdx];
+          const auraXml = `${auraPath}/${auraName}/${auraName}.cmp-meta.xml`;
+          let auraJSON = parseXML(auraXml);
+          loggit.logLine('Checking Aura Component ' + auraName);
+          if (auraJSON['AuraDefinitionBundle'] && auraJSON['AuraDefinitionBundle']['apiVersion']) {
+            setValue(apiVersions,`AuraDefinitionBundle.${auraName}`,parseFloat(auraJSON['AuraDefinitionBundle']['apiVersion'][0]));
+           /* if (apiVersions['AuraDefinitionBundle'] == undefined) {
+              apiVersions['AuraDefinitionBundle'] = {};
+            }
+            apiVersions['AuraDefinitionBundle'][auraName] = parseFloat(auraJSON['AuraDefinitionBundle']['apiVersion'][0]);*/
+          }
+          //Count Used Components by Namespace
+          let auraCmpFile = `${auraPath}/${auraName}/${auraName}.cmp`;
+          loggit.logLine(`Extracting info from ${auraCmpFile}`);
+          //TODO: Parse for object references
+
+          if (fs.existsSync(auraCmpFile)) {
+
+            let auraBody = fs.readFileSync(auraCmpFile, 'utf8');
+
+            loggit.logLine('Performing Regex search against component for namespaces');
+            const componentsReg = /<(\w+:\w+)/ig;
+            let referencedComponents = getMatches(auraBody, componentsReg);
+            if (referencedComponents.length > 0) {
+              loggit.logLine(`Found the following Components: ${JSON.stringify(referencedComponents)}`);
+           //   addKeyToObject(componentProperties,`AuraDefinitionBundle.${auraName}.namespaceReferences`);
+          /*    if (componentProperties['AuraDefinitionBundle'] == undefined) {
+                componentProperties['AuraDefinitionBundle'] = {};
+              }
+              if (componentProperties['AuraDefinitionBundle'][auraName] == undefined) {
+                componentProperties['AuraDefinitionBundle'][auraName] = {};
+              }
+              if (componentProperties['AuraDefinitionBundle'][auraName]['namespaceReferences'] == undefined) {
+                componentProperties['AuraDefinitionBundle'][auraName]['namespaceReferences'] = {};
+              }*/
+              referencedComponents.forEach(element => {
+                let ns = element.split(":", 2)[0];
+                incrementValue(componentProperties,`AuraDefinitionBundle.${auraName}.namespaceReferences.${ns}`);
+              /*  if (componentProperties['AuraDefinitionBundle'][auraName]['namespaceReferences'][ns] == undefined) {
+                  componentProperties['AuraDefinitionBundle'][auraName]['namespaceReferences'][ns] = 1;
+                }
+                else {
+                  componentProperties['AuraDefinitionBundle'][auraName]['namespaceReferences'][ns] += 1
+                }*/
+                //Also add it to the Namespaces dependencies
+                setValue(dependencies,`namespaces.${ns}`,1);
+             /*   if (dependencies['namespaces'] == undefined) {
+                  dependencies['namespaces'] = {};
+                }
+                dependencies['namespaces'][ns] = 1;*/
+              });
+            }
+            loggit.logLine('Extracting implemented and extended interfaces');
+            const interfaceReg = /(?:implements|extends)\s*=\s*"([\w ,:]+)"/igm;
+            let interfaceMatches = getMatches(auraBody, interfaceReg);
+            if (interfaceMatches.length > 0) {
+              loggit.logLine(`Found the following Interfaces: ${JSON.stringify(interfaceMatches)}`);
+          /*    if (componentProperties['AuraDefinitionBundle'] == undefined) {
+                componentProperties['AuraDefinitionBundle'] = {};
+              }
+              if (componentProperties['AuraDefinitionBundle'][auraName] == undefined) {
+                componentProperties['AuraDefinitionBundle'][auraName] = {};
+              }
+              if (componentProperties['AuraDefinitionBundle'][auraName]['interfaces'] == undefined) {
+                componentProperties['AuraDefinitionBundle'][auraName]['interfaces'] = {};
+              }*/
+              interfaceMatches.forEach(element => {
+                let interfaces = element.split(/ *, */);
+                interfaces.forEach(element => {
+             //     componentProperties['AuraDefinitionBundle'][auraName]['interfaces'][element] = 1;
+                  setValue(componentProperties,`AuraDefinitionBundle.${auraName}.interfaces.${element}`,1);
+                });
+              })
+            }
+            //Find Object References
+            addObjectDependencies(dependencies,extractObjectsApex(auraBody));
+        /*    const findObjectsReg = /(?:(?<namespace>[a-zA-Z](?:[a-z]|[A-Z]|[0-9]|_(?!_)){0,14})__)?(?<component>(?<!___)[a-zA-Z](?:[a-z]|[A-Z]|[0-9]|_(?!_))+)(?:__(?<suffix>c|mdt|e|x|b|pc|pr|r|xo|latitude__s|longitude__s|history|ka|kav|feed|share))/g;
+            let objectFound;
+            loggit.logLine('Looking for fields and objects referenced within Aura Components')
+            while (objectFound = findObjectsReg.exec(auraBody)) {
+              loggit.logLine('Found: ' + objectFound[0]);
+              const object = getNameSpaceAndType(objectFound[0]);
+              //Add Namespace Dependencies 
+              if (object.namespace !== null) {
+                if (dependencies['namespaces'] == undefined) {
+                dependencies['namespaces'] = {};
+                }
+                dependencies['namespaces'][object.namespace] = 1;
+              }
+              if (dependencies['components'] == undefined) {
+                dependencies['components'] = {};
+              }
+              dependencies['components'][object.fullName] = object;
+            } */  
+
+          }
+          else {
+            loggit.logLine('File not found');
+          }
+        }
+        break;
+    }
+
+    inventory[metadataType] = typeInv;
+   // this.ux.stopSpinner('Complete');
+  }
+ 
+  //Check Person Accounts
+  let pafile = `${sourceDir}/objects/PersonAccount.object`;
+  if (fs.existsSync(pafile)) {
+    setValue(dependencies,'features.PersonAccount',1);
+   /* if (dependencies['features'] == undefined) {
+      dependencies['features'] = {};
+    }
+    dependencies['features']['PersonAccount'] = 1;*/
+  }
+
+  inventory['apiVersions'] = apiVersions;
+  inventory['componentProperties'] = componentProperties;
+  inventory['dependencies'] = dependencies;
+  return inventory;
+}
+
+function getMembers(mdTypeDef, sourceDir) {
+//  this.loggit.logLine('Getting wildcard members for ' + mdTypeDef.name);
+  let retVal = mdTypeDef['members'];
+  if (mdmap[mdTypeDef.name] != undefined) {
+    if (mdmap[mdTypeDef.name]['folder'] != 'null' && mdmap[mdTypeDef.name]['extension'] != 'null') {
+      retVal = getMembersFromFiles(`${sourceDir}/${mdmap[mdTypeDef.name]['folder']}`, mdmap[mdTypeDef.name]['extension']);
+      //        this.loggit.loggit("Added Members from files.:" + JSON.stringify(retVal));
+    }
+  }
+  return retVal;
+}
+
+
+function getMembersFromFiles(folder, extension) {
+  const typePath = folder;
+  const members = [];
+  if (!fs.existsSync(typePath)) {
+//    this.loggit.logLine(`Folder ${typePath} does not exist. Cannot find members`);
+    return members;
+  }
+//  this.loggit.logLine(`Looking in folder ${typePath} for members`);
+  const folderContents = fs.readdirSync(typePath);
+//  this.loggit.logLine('Folder Contents: ' + JSON.stringify(folderContents));
+  folderContents.forEach(element => {
+
+    const [fileName, ext] = [element.substr(0, element.lastIndexOf('.')), element.substr(element.lastIndexOf('.') + 1, element.length)]
+    if (ext === extension) {
+      members.push(fileName);
+    }
+  });
+
+//  this.loggit.logLine('Found Members: ' + JSON.stringify(members));
+  return members;
+}
+
+function extractObjectsApex(apexBody: string)  {
+  const findObjectsReg = /(?:(?<namespace>[a-zA-Z](?:[a-z]|[A-Z]|[0-9]|_(?!_)){0,14})__)?(?<component>(?<!___)[a-zA-Z](?:[a-z]|[A-Z]|[0-9]|_(?!_))+)(?:__(?<suffix>c|mdt|e|x|b|pc|pr|r|xo|latitude__s|longitude__s|history|ka|kav|feed|share))/g;
+   
+  let objectsFound = [];
+  let objectFound;
+  while (objectFound = findObjectsReg.exec(apexBody)) {
+    objectsFound.push(getNameSpaceAndType(objectFound[0]));
+  }  
+  return objectsFound;      
+ }
+
+ function addObjectDependencies(dependencies: Object, objects: Object[]) {
+ 
+    for (const o of objects) {
+      if (o['namespace'] !== null) {
+        setValue(dependencies,`namespaces.${o['namespace']}`,1);
+      }
+      setValue(dependencies,`components.${o['fullName']}`,o);
+    }
+ }
+
+
+
+function stripApexComments(apexBody : string) {
+  const commentReg = /\/\*[\s\S]*?\*\/|((?<!:)|^)\/\/.*$/gm;
+  return apexBody.replace(commentReg,'');
+}
+
+
+function getMatches(searchString, regex) {
+  let matches = [];
+  let match;
+  while (match = regex.exec(searchString)) {
+    matches.push(match[1]);
+  }
+ // logLine(`Found ${matches.length} matches`);
+ // logJSON(matches);
+  return matches;
+}
+
+function getNameSpaceAndType(fullComponentName) {
+  /*
+  Parse object or field names
+  Cases Covered:
+  StandardThing
+  StandardThing__ext
+  CustomThing__ext
+  Namespace_CustomThing__ext
+  CustomField__location__s
+  Namespace__CustomField__location__s
+  */
+ // logLine('Breaking down component name:' + fullComponentName);
+  
+  const retVal = {
+    type:null,
+    name:null,
+    extension:null,
+    namespace:null,
+    fullName: fullComponentName
+  };
+
+  let breakdown = fullComponentName.split("__");
+ // logLine('Component Breakdown: ' + JSON.stringify(breakdown));
+  if (breakdown.length == 1) {
+      //Standard Object
+      retVal.type = 'Standard';
+      retVal.name = breakdown[0];
+  }
+  else {
+    //Check for and fixup locations
+ //   logLine('Proposed Suffix:' + breakdown[breakdown.length -1]);
+    if (breakdown[breakdown.length -1] == 's'  && breakdown.length > 2 && (breakdown[breakdown.length-2] == 'latitude' || breakdown[breakdown.length-2] == 'longitude' )) {
+   //   logLine('This appears to be a location field with an extra __ in the suffix. Adjusting expectations'); 
+        breakdown.pop();
+        breakdown[breakdown.length -1] = breakdown[breakdown.length -1] + '__s';
+      
+    }
+
+    //we have an extension and possibly a namespace
+    retVal.extension = breakdown[breakdown.length -1];
+    
+    if (breakdown.length == 2) {
+      //non namespaced custom object
+      retVal.name = breakdown[0];
+    }
+    else if (breakdown.length == 3) {
+      retVal.name = breakdown[1];
+      retVal.namespace = breakdown[0];
+    }
+    else {
+      //WE probably shouldn't end up here.
+  //    logLine('Unsure how to parse ' + fullComponentName);
+      retVal.name = fullComponentName;
+    }
+    switch (retVal.extension) {
+      case 'c':
+        retVal.type = 'Custom';
+        break;
+      case 'r':
+        retVal.type = 'Relationship';
+        break;
+      case 'ka':
+        retVal.type = 'Knowledge Article';
+        break;
+      case 'kav':
+        retVal.type = 'Knowledge Article Version';
+        break;
+      case 'feed':
+        retVal.type = 'Object Feed';
+        break;
+      case 'viewstat':
+        retVal.type = 'Knowledge Article View Stat';
+        break;
+      case 'votestat':
+        retVal.type = 'Knowledge Article Vote Stat';
+        break;
+      case 'datacategoryselection':
+        retVal.type = 'Knowledge Article Data Category';
+        break;
+      case 'x':
+        retVal.type = 'External Object';
+        break;
+      case 'xo':
+        retVal.type = 'S2S Proxy Object';
+        break;
+      case 'mdt':
+        retVal.type = 'Custom Metadata Type';
+        break;
+      case 'share':
+        retVal.type = 'Custom Object Sharing';
+        break;
+      case 'tag':
+        retVal.type = 'Tag';
+        break;
+      case 'history':
+        retVal.type = 'Field History Tracking';
+        break;
+      case 'pc':
+        retVal.type = 'Person Account';
+        break;
+      case 'pr':
+        retVal.type = 'Person Account Relationship';
+        break;
+      case 'b':
+        retVal.type = 'Big Object';
+        break;
+      case 'latitude__s':
+        retVal.type = 'Geolocation Latitude Coordinate';
+        break;
+      case 'longitude__s':
+        retVal.type = 'Geolocation Longitude Coordinate';
+        break;
+      case 'e':
+        retVal.type = 'Platform Event';
+        break;
+      case 'p':
+        retVal.type = 'Custom Person Object';
+        break;
+      case 'changeevent':
+        retVal.type = 'Change Data Capture';
+        break;
+      default:
+        retVal.type = 'Unknown';
+    }
+
+  }
+//  logLine('Parsed component: ' + JSON.stringify(retVal));
+  return retVal;
+}
+
+export function parseXML(xmlfile, dieOnError = false) {
+  const parser = new xml2js.Parser({
+    attrkey: 'ATTR'
+  });
+  let json = [];
+  let error = undefined;
+
+  if (!fs.existsSync(xmlfile)) {
+    let message = `Cannot find XML File: ${xmlfile}`;
+    if (dieOnError) {
+ //     this.loggit.logLine(message, 'Error');
+      throw new SfdxError(message, 'XMLNotFoundError');
+    } else {
+ //     this.loggit.logLine(message, 'Warn');
+      return json;
+    }
+  }
+
+  let xmlData = fs.readFileSync(xmlfile, 'utf8');
+  parser.parseString(xmlData.substring(0, xmlData.length), function (err, result) {
+    error = err;
+    json = result;
+  });
+
+  if (error) {
+ //   this.loggit.logLine(`Error parsing ${xmlfile}: ${error}`, 'Error');
+    throw new SfdxError(`Error parsing ${xmlfile}: ${error}`, 'XMLParseError');
+  }
+  return json;
+}
+
+
+
+
+
+
+
